simple-cdd-ext: Add helper to create shell wrapper on any executable
It might be necessary to catch specific invocation (i.e. with specific
arguments) of binary in installer environment to perform specific
actions before (pre hook) and after (post hook) execution.

For example in debian-installer we might want to catch udpkg calls to
patch postinst scripts after --unpack operations. To do this post
hook will be used that checks for unpack option with specific package
to consider if postinst script patch should be applied.

Signed-off-by: Serhey Popovych <serhe.popovych@gmail.com>